+++ lib/controllers/tampil seluruh data/all_global_controller.dart
@@ -1,17 +1,23 @@
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:get/get.dart';
 
+import '../../helpers/connectivity.dart';
 import '../../models/tampil seluruh data/do_global_all.dart';
 import '../../models/user_model.dart';
 import '../../repository/tampil seluruh data/global_all_repo.dart';
 import '../../utils/constant/storage_util.dart';
 import '../../utils/popups/dialogs.dart';
 import '../../utils/popups/full_screen_loader.dart';
+import '../../utils/popups/snackbar.dart';
 
 class DataAllGlobalController extends GetxController {
   final isLoadingGlobalHarian = Rx<bool>(false);
   RxList<DoGlobalAllModel> doAllGlobalModel = <DoGlobalAllModel>[].obs;
   final dataGlobalAllRepo = Get.put(GlobalAllRepository());
   final storageUtil = StorageUtil();
+
+  final isConnected = Rx<bool>(true);
+  final networkManager = Get.find<NetworkManager>();
 
   // roles users
   int rolesEdit = 0;
@@ -26,13 +32,45 @@ class DataAllGlobalController extends GetxController {
       rolesEdit = user.edit;
       rolesHapus = user.hapus;
     }
+
+    // Listener untuk memantau perubahan koneksi
+    networkManager.connectionStream.listen((ConnectivityResult result) {
+      if (result == ConnectivityResult.none) {
+        // Jika koneksi hilang, tampilkan pesan
+        if (isConnected.value) {
+          SnackbarLoader.errorSnackBar(
+            title: 'Koneksi Terputus',
+            message: 'Anda telah kehilangan koneksi internet.',
+          );
+          isConnected.value = false;
+        }
+      } else {
+        // Jika koneksi kembali, perbarui status koneksi
+        if (!isConnected.value) {
+          isConnected.value = true;
+          fetchAllGlobalData(); // Otomatis fetch data ketika koneksi kembali
+        }
+      }
+    });
+
     fetchAllGlobalData();
     super.onInit();
   }
 
   Future<void> fetchAllGlobalData({DateTime? pickDate}) async {
     try {
+      final connectionStatus = await networkManager.isConnected();
+      if (!connectionStatus) {
+        isConnected.value = false;
+        SnackbarLoader.errorSnackBar(
+          title: 'Tidak ada koneksi internet',
+          message: 'Silakan coba lagi setelah koneksi tersedia',
+        );
+        return;
+      }
+
       isLoadingGlobalHarian.value = true;
+      isConnected.value = true;
       final dataHarian = await dataGlobalAllRepo.fetchGlobalHarianContent();
 
       if (pickDate != null) {
@@ -71,6 +109,15 @@ class DataAllGlobalController extends GetxController {
   ) async {
     CustomDialogs.loadingIndicator();
 
+    final isConnected = await networkManager.isConnected();
+    if (!isConnected) {
+      CustomFullScreenLoader.stopLoading();
+      SnackbarLoader.errorSnackBar(
+          title: 'Tidak ada koneksi internet',
+          message: 'Silahkan coba lagi setelah koneksi tersedia');
+      return;
+    }
+
     await dataGlobalAllRepo.editDOGlobalContent(
         id, tgl, idPlant, tujuan, srd, mks, ptk, bjm);
     CustomFullScreenLoader.stopLoading();
@@ -83,6 +130,15 @@ class DataAllGlobalController extends GetxController {
     int id,
   ) async {
     CustomDialogs.loadingIndicator();
+
+    final isConnected = await networkManager.isConnected();
+    if (!isConnected) {
+      CustomFullScreenLoader.stopLoading();
+      SnackbarLoader.errorSnackBar(
+          title: 'Tidak ada koneksi internet',
+          message: 'Silahkan coba lagi setelah koneksi tersedia');
+      return;
+    }
     await dataGlobalAllRepo.deleteDOGlobalContent(id);
 
     await fetchAllGlobalData();
